syntax errors, still trying to fix vercel deploy
