+++ package.json
@@ -9,22 +9,21 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/node": "20.10.6",
-    "@types/react": "18.2.46",
+    "@types/node": "20.11.5",
+    "@types/react": "18.2.48",
     "@types/react-dom": "18.2.18",
-    "autoprefixer": "10.4.16",
-    "axios": "^1.6.3",
+    "autoprefixer": "10.4.17",
+    "axios": "^1.6.5",
     "bcryptjs": "^2.4.3",
-    "next": "14.0.4",
+    "next": "14.1.4",
     "next-auth": "^4.24.5",
-    "postcss": "8.4.32",
+    "postcss": "8.4.33",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "3.4.0",
+    "tailwindcss": "3.4.1",
     "typescript": "5.3.3"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.6",
-    "eslint": "^9.15.0"
+    "@types/bcryptjs": "^2.4.6"
   }
 }
